Fix camera aspect ratio

+++ DemoParticles/Src/Camera/CameraControllerFPS.cpp
@@ -14,7 +14,7 @@ namespace DemoParticles
     {
         float yaw = 0.0f;
         float pitch = 0.0f;
-        m_camera = std::make_unique<Camera>(Vector3(0.0f, 0.0f, 0.0f), yaw, pitch, 0.0f, m_deviceResources->GetOutputWidth() / m_deviceResources->GetOutputHeight(), DirectX::XM_PI / 4.0f);
+        m_camera = std::make_unique<Camera>(Vector3(0.0f, 0.0f, 0.0f), yaw, pitch, 0.0f, (float)m_deviceResources->GetOutputWidth() / m_deviceResources->GetOutputHeight(), DirectX::XM_PI / 4.0f);
     }
 
     //needed to use forward declaration of a unique_ptr :
